JSON serialization refactoring - work in progress.
